Upade - synch to 4.2 Dev - Vertex Color force is made as a choice
Vertex Color Shade node use is currently forced on the user - this caused render issues and sim issues on aircraft. This is now a choice - default off, for the user in the install preferences.
Code clean up
Parallax - bug fix for errors
SSS - fix errors
Default rougness and metallic back to original ASOBO defaults - as they are more in line with Substance Painter.